backend/requirements.txt
django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
python-dotenv==1.0.1
firebase-admin==6.6.0
google-cloud-aiplatform==1.80.0
google-generativeai==0.3.2 